Fixed controllers, added LoginMethodMiddleware
