+++ app/Http/Controllers/Controller.php
@@ -16,45 +16,17 @@ class Controller extends BaseController
 {
     use AuthorizesRequests, AuthorizesResources, DispatchesJobs, ValidatesRequests;
 
-    protected function isOauthDefined() {
-    	$oAuth = config('oauth.oAuthProvider');
-        $oAuthDefined = false;
-        if(!empty($oAuth)) {
-            $oAuthDefined = true;
-        }
-
-        return $oAuthDefined;
-    }
-
-    protected function getOAuthProvider() {
-    	$oAuth = config('oauth.oAuthProvider');
-        return $oAuth;
-    }
-
-    protected function getDelegatedAdmin() {
-        return config('oauth.oAuthAdmin');
-    }
-
-    protected function getOAuthAllowedDomain() {
-    	$oAuthDomain = config('oauth.oAuthDomain');
-    	return $oAuthDomain;
-    }
-
-    protected function getOauthCredentials($adminId = '') {
-        if($this->getOAuthProvider() == 'google') {
-            $path = storage_path()."/app/".config('oauth.credentials').".json";
-            return json_decode(File::get($path), true);
-        } elseif($this->getOAuthProvider() == 'azure') {
-            $user = User::findOrFail($adminId);
-            $toReturn = array(
-                'token' =>  $user->oauth_token
-            );
-            return $toReturn;
-        }
-    }
-
     protected function getAppVersion() {
         $path = storage_path()."/../.version";
         return File::get($path);
+    }
+
+    protected function getUpdateArray($req, $fields, $allowNull = true) {
+        $arr = array();
+
+        foreach($fields as $field) {
+            if ($req->has($field) && ($allowNull || $req->get($field)!=null)) { $arr[$field] = $req->get($field);}
+        }
+        return $arr;
     }
 }
